requirements.txt
flask==3.0.0
python-dotenv==1.0.0
groq==0.4.1
google-generativeai==0.3.2
supabase==2.8.1
httpx==0.27.0
requests==2.31.0
gunicorn==21.2.0
colorama==0.4.6
